+++ packages/design-system/hooks/useLogoOverrides/constants.ts
@@ -1,0 +1,8 @@
+import { LogoOverrides } from './types'
+
+export const DEFAULT_LOGO_KEY = 'logo-overrides'
+
+export const DEFAULT_LOGO_SETTINGS: LogoOverrides = {
+  default: undefined,
+  square: undefined,
+}